update id to long, update entity relation

+++ src/main/java/id/co/alamisharia/repository/ProductRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
-    List<Product> findBySellerId(int sellerId);
+public interface ProductRepository extends JpaRepository<Product, Long> {
+    List<Product> findBySellerId(Long sellerId);
 
     List<Product> findByNamaContains(String keyword);
