+++ src/main/java/id/co/alamisharia/repository/SellerRepository.java
@@ -3,5 +3,5 @@ package id.co.alamisharia.repository;
 import id.co.alamisharia.entity.Seller;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SellerRepository extends JpaRepository<Seller, Integer> {
+public interface SellerRepository extends JpaRepository<Seller, Long> {
 }
